LayoutTests/storage/domstorage/sessionstorage/set-item-synchronous-keydown-expected.txt

Tests updating sessionStorage in the keydown handler and makes sure the value is updated synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sessionStorage.setItemSynchronousKeydownTestValue is "0"
Setting setItemSynchronousKeydownTestValue to 1 in keydown handler.
PASS sessionStorage.setItemSynchronousKeydownTestValue is "1"
PASS successfullyParsed is true

TEST COMPLETE
